draft pruning after symmetrisation
